// src/app/models/category.enum.ts
export enum CategoryEnum {
  Plastic = 'Plásticos',
  BrownGlass = 'Vidrios marrones',
  Paper = 'Papeles',
  GreenGlass = 'Vidrios verdes',
  WhiteGlass = 'Vidrios blancos',
  Metal = 'Metales',
  Battery = 'Baterias',
  Biological = 'Biológicos',
  Trash = 'Basura',
  Cardboard = 'Cartones',
  Shoes = 'Zapatos',
  Clothes = 'Ropas'
}